requirements.txt
Django>=1.9,<1.10
dj-database-url==0.4.1
django-dotenv==1.4.1
django-markupfield==1.4.0
Markdown==2.6.6
docutils==0.9
pep8==0.6.1
psycopg2==2.4.5
pycrypto==2.6
pyflakes==0.5.0
pytz
requests==2.5.0
stripe==1.37.0
